공지사항 벡터화 및 파일 경로 변경
- OCR이 적용된 공지사항 벡터화 완료
- 목표 달성까지 무한 반복 크롤링 기능 추가
- 파일 경로 개선 -> 실행 파일 전부 최상위 경로로 위치
