+++ ai/.gitignore
@@ -2,3 +2,4 @@
 
 .env
 *.pkl
+*.zip
