+++ ai/crawler/clova_ocr.py
@@ -1,5 +1,3 @@
-# OCR 클래스
-
 import requests
 import json
 import time
@@ -25,26 +23,21 @@ class OCR:
 
   def img_to_txt(self, img_url):
     headers = {'User-Agent' : 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/123.0.0.0 Safari/537.36'}
-    filename = "tmp_image.jpg"
-    time.sleep(random.uniform(2,5))
+    filename = "./tmp_image.jpg"
+    #time.sleep(random.uniform(2,3))
     resp = requests.get(img_url, headers=headers, verify=False)
-
     if resp.status_code != 200:
       return ''
+    
     with open(filename, "wb") as f:
         f.write(resp.content)
     files = [('file', open('./tmp_image.jpg','rb'))]
-    time.sleep(random.uniform(2,5))
+    #time.sleep(random.uniform(2,3))
     response = requests.request("POST", self.api_url, headers=self.headers, data=self.payload, files=files)
     res = response.json()
-
-    if 'images' not in res.keys():
-      return ''
-
     return self.join_text(res)
 
   def img_lst_to_txt(self, img_url_lst):
-
     if not img_url_lst:
       return
 
@@ -58,7 +51,7 @@ class OCR:
 
   def join_text(self, result):
     txt = ''
-    if 'fields' not in result['images'][0].keys():
+    if 'images' not in result.keys() or 'fields' not in result['images'][0].keys():
       return txt
     for img in result['images'][0]['fields']:
       txt += img['inferText']
